+++ extensions/autumn/src/manage-customers.tsx
@@ -1,18 +1,23 @@
-import { Action, ActionPanel, Form, Icon, List, showToast, Toast, useNavigation } from "@raycast/api";
-import { useCachedPromise, useForm } from "@raycast/utils";
+import { Action, ActionPanel, Alert, confirmAlert, Detail, Icon, Keyboard, List, showToast, Toast } from "@raycast/api";
+import { useCachedPromise } from "@raycast/utils";
 import { autumn, AUTUMN_LIMIT } from "./autumn";
+import CreateCustomer from "./components/create-customer";
 
 export default function ManageCustomers() {
   const {
     isLoading,
     data: customers,
     error,
-    revalidate,
+    mutate,
+    pagination,
   } = useCachedPromise(
-    async () => {
-      const { data, error } = await autumn.customers.list({ limit: AUTUMN_LIMIT });
+    () => async (options) => {
+      const { data, error } = await autumn.customers.list({ limit: AUTUMN_LIMIT, offset: options.page*AUTUMN_LIMIT});
       if (error) throw new Error(error.message);
-      return data.list;
+      return {
+        data: data.list,
+        hasMore: data.total===data.limit
+      }
     },
     [],
     {
@@ -21,8 +26,8 @@ export default function ManageCustomers() {
   );
 
   return (
-    <List isLoading={isLoading}>
-      {!isLoading && !customers.length && error ? (
+    <List isLoading={isLoading} pagination={pagination}>
+      {!isLoading && !customers.length && !error ? (
         <List.EmptyView
           description="Create your first customer by interacting with an Autumn function via the API."
           actions={
@@ -30,8 +35,7 @@ export default function ManageCustomers() {
               <Action.Push
                 icon={Icon.AddPerson}
                 title="Create Customer"
-                target={<CreateCustomer />}
-                onPop={revalidate}
+                target={<CreateCustomer onCreate={mutate} />}
               />
             </ActionPanel>
           }
@@ -52,12 +56,55 @@ export default function ManageCustomers() {
             ]}
             actions={
               <ActionPanel>
+                {customer.id && <Action.Push icon={Icon.Person} title="Customer Details" target={<CustomerDetails customerId={customer.id as string} />} />}
                 <Action.Push
                   icon={Icon.AddPerson}
                   title="Create Customer"
-                  target={<CreateCustomer />}
-                  onPop={revalidate}
+                  target={<CreateCustomer onCreate={mutate} />}
                 />
+                {customer.id && <Action icon={Icon.RemovePerson} title="Delete Customer" onAction={() => confirmAlert({
+                  title: "Delete Customer",
+                  message: "Are you sure you want to delete this customer in Autumn? This action cannot be undone.",
+                  primaryAction: {
+                    style: Alert.ActionStyle.Destructive,
+                    title: "Delete",
+                    async onAction() {
+                      const delete_in_stripe = await confirmAlert({
+                        title: "Select whether to delete this customer in Stripe as well.",
+                        dismissAction: {
+                          title: "Delete in Autumn only"
+                        },
+                        primaryAction: {
+                          style: Alert.ActionStyle.Destructive,
+                          title: "Delete in Autumn and Stripe"
+                        },
+                        rememberUserChoice: true
+                      });
+                        const customerId = customer.id as string;
+                        const toast = await showToast(Toast.Style.Animated, "Deleting", customerId);
+                        try {
+                          await mutate(
+                            autumn.customers.delete(customerId, {
+                              delete_in_stripe
+                            }).then(({error}) => {
+                                if (error) throw new Error(error.message);
+                            }), {
+                              optimisticUpdate(data) {
+                                return data.filter(c => c.id!==customer.id)
+                              },
+                              shouldRevalidateAfter: false
+                            }
+                          )
+                          toast.style = Toast.Style.Success;
+                          toast.title = "Deleted";
+                        } catch (error) {
+                          toast.style = Toast.Style.Failure;
+                          toast.title = "Failed";
+                          toast.message = `${error}`;
+                        }
+                    }
+                  }
+                })} style={Action.Style.Destructive} shortcut={Keyboard.Shortcut.Common.Remove} />}
               </ActionPanel>
             }
           />
@@ -67,48 +114,17 @@ export default function ManageCustomers() {
   );
 }
 
-function CreateCustomer() {
-  const { pop } = useNavigation();
-  type CreateCustomer = {
-    name: string;
-    id: string;
-    email: string;
-  };
-  const { handleSubmit, itemProps, values } = useForm<CreateCustomer>({
-    async onSubmit(values) {
-      const toast = await showToast(Toast.Style.Animated, "Creating", values.name || values.email || values.id);
-      try {
-        const { error } = await autumn.customers.create(values);
-        if (error) throw new Error(error.message);
-        toast.style = Toast.Style.Success;
-        toast.title = "Created";
-        pop();
-      } catch (error) {
-        toast.style = Toast.Style.Failure;
-        toast.title = "Failed";
-        toast.message = `${error}`;
-      }
-    },
-    validation: {
-      id(value) {
-        if (!value && !values.email) return "ID or email is required";
-      },
-      email(value) {
-        if (!value && !values.id) return "ID or email is required";
-      },
-    },
-  });
-  return (
-    <Form
-      actions={
-        <ActionPanel>
-          <Action.SubmitForm icon={Icon.AddPerson} title="Create Customer" onSubmit={handleSubmit} />
-        </ActionPanel>
-      }
-    >
-      <Form.TextField title="Name" {...itemProps.name} />
-      <Form.TextField title="ID" {...itemProps.id} info="Your unique identifier for the customer" />
-      <Form.TextField title="Email" {...itemProps.email} />
-    </Form>
-  );
+function CustomerDetails({customerId}:{customerId:string}) {
+  const {isLoading, data: customer} = useCachedPromise(async(id: string) => {
+    const {data,error} = await autumn.customers.get(id);
+    if (error) throw new Error(error.message);
+    return data;
+  }, [customerId])
+  return <Detail isLoading={isLoading} metadata={customer&&<Detail.Metadata>
+<Detail.Metadata.Label title="ID" text={customer.id || "N/A"} />
+<Detail.Metadata.Label title="Name" text={customer.name || "None"} />
+<Detail.Metadata.Label title="Email" text={customer.email || "None"} />
+<Detail.Metadata.Label title="Fingerprint" text={customer.fingerprint || "None"} />
+  </Detail.Metadata>} />
 }
+