+++ extensions/autumn/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Autumn Changelog
 
+## [] - {PR_MERGE_DATE}
+
+- Delete Product
+- Add "Open in Autumn" component
+- List Features
+- List Customers is now paginated
+- Delete Customers
+- Customer Details
+
 ## [Initial Version] - 2025-09-15
 
 - List Customers